freebsd: always prefer flsl*() for 14+
FreeBSD has used HAVE_INLINE_FLS* macros to indicate presence of a
"fast" implementation of find-last-set type functions. On platforms
where the inline implementation is unavailable, some functions have
provided an alternate implementation that performs better than the fls*
functions provided by the kernel.

A recent commit to FreeBSD [1] switched the ffs/fls functions to always
use what is provided by the compiler, with the basic justification that
the compiler runtime routines are well-optimized enough to rely on
generally.

The result is that the HAVE_INLINE_F** macros are now unconditionally
defined. These macros and the few related hand-crafted optimizations
are being removed.

The highbit() and highbit64() functions in OpenZFS have a fallback
binary search implementation when the relevant inline function is not
available. FreeBSD 14 and onward, always use flsl()/flsll().

[1] https://cgit.freebsd.org/src/commit/?id=a89262079edba80ba6133e81af1e14ab7386f53a